fix: harden cursor and crosshair integration
