+++ docs/guides/custom-marks-and-renderers.md
@@ -181,6 +181,8 @@ A mark that emits only cursor-driven rules, labels, or markers declares that
 role explicitly:
 
 ```ts
+import { resolveCrosshairGuide } from '@tanstack/charts/crosshair'
+
 return {
   id,
   channels: {},
@@ -201,6 +203,7 @@ return {
             const position = scale.map(value)
             return Number.isFinite(position) ? position : undefined
           },
+          resolve: resolveCrosshairGuide,
         },
       ],
     }
@@ -217,9 +220,14 @@ do not need to invent a placement for an ordinary guide mark.
 
 The required `surface` bounds cover the complete chart surface; `chart` covers
 the inner plot. Use `chart` for clipped rules and `surface` for labels that must
-remain visible. A custom renderer receives final `SceneFocusGuide` values after
-the compiler has filled in placement. Pass the scene to
-`resolveFocusPresentation` instead of resolving mark order inside the renderer.
+remain visible. Each guide's required `resolve` callback receives the final
+guide, local focus, pointer, and projected cursor, then returns one transient
+scene node or `undefined`. `resolveCrosshairGuide` provides the built-in rule,
+band, label, and marker behavior. A custom guide can supply different policy
+without adding it to renderer bundles that never use the guide. A custom
+renderer receives final `SceneFocusGuide` values after the compiler has filled
+in placement. Pass the scene to `resolveFocusPresentation` instead of calling
+guide resolvers or resolving mark order inside the renderer.
 
 ## Separate point and scale values
