+++ packages/charts-core/src/cursor-host.ts
@@ -1,9 +1,17 @@
 export {
+  cursorHost,
   createFocusChartCursorState,
   createFreeChartCursorState,
   resolveChartCursorFocus,
   resolveChartCursorPresentation,
+} from './cursor'
+export {
   resolveChartFocusStrategy,
   resolveChartPointerFocus,
-} from './cursor'
+} from './interaction'
+export { createChartCursorHostSession } from './cursor-host-contract'
+export type {
+  ChartCursorHostExtension,
+  ChartCursorHostSession,
+} from './cursor-host-contract'
 export { resolveFocusPresentation } from './focus-presentation'
