+++ packages/charts-core/src/cursor-public.ts
@@ -1,7 +1,8 @@
-export { createChartCursor } from './cursor'
+export { createChartCursor, cursorHost } from './cursor'
 export type {
   ChartCursorController,
   ChartCursorCoordinates,
+  ChartCursorExtensionToken,
   ChartCursorPointIdentity,
   ChartCursorState,
   ChartCursorStateUpdater,
