+++ packages/charts-core/src/cursor.ts
@@ -1,6 +1,13 @@
-import { focusNearestX, focusNearestY, focusX, focusY } from './focus'
-import { findContainingScenePoint } from './nearest'
+import { type ChartCursorHostExtension } from './cursor-host-contract'
 import { valueKey } from './scales'
+export {
+  chartPointFromNavigationOrder,
+  chartPointFromSceneOrder,
+  resolveChartFocusStrategy,
+  resolveChartPointerFocus,
+  restoreChartFocusPoint,
+  sameChartPointIdentity,
+} from './interaction'
 import type {
   ChartCursorBinding,
   ChartCursorController,
@@ -8,9 +15,7 @@ import type {
   ChartCursorPointIdentity,
   ChartCursorPresentation,
   ChartCursorState,
-  ChartCursorStateUpdater,
   ChartCursorValues,
-  ChartFocusMode,
   ChartFocusSource,
   ChartFocusState,
   ChartFocusStrategy,
@@ -47,6 +52,62 @@ export function createChartCursor<
       for (const listener of [...listeners]) listener()
     },
   }
+}
+
+/** Platform-neutral cursor behavior injected into chart hosts by a binding. */
+export const cursorHost: ChartCursorHostExtension = {
+  id: 'cursor',
+  __chartExtensionType: 'cursor',
+  create(controller) {
+    let publishedState: ReturnType<typeof controller.getState> | undefined
+    let publishing = false
+    return {
+      controller,
+      getState: () => controller.getState(),
+      subscribe: (listener) =>
+        controller.subscribe(() => {
+          if (publishing) {
+            publishedState = controller.getState() ?? undefined
+          }
+          listener()
+        }),
+      owns: (state) => state !== null && state === publishedState,
+      publish(state) {
+        publishedState = state
+        publishing = true
+        try {
+          controller.setState(state)
+        } finally {
+          publishedState = controller.getState() ?? undefined
+          publishing = false
+        }
+      },
+      clearOwnedTransient() {
+        const current = controller.getState()
+        if (!current || current !== publishedState || current.pinned) {
+          return false
+        }
+        publishedState = undefined
+        controller.setState(null)
+        return true
+      },
+      clear() {
+        publishedState = undefined
+        controller.setState(null)
+      },
+      destroy() {
+        const current = controller.getState()
+        if (current && current === publishedState && !current.pinned) {
+          controller.setState(null)
+        }
+        publishedState = undefined
+      },
+      resolvePresentation: resolveChartCursorPresentation,
+      resolveFocus: resolveChartCursorFocus,
+      createFocusState: createFocusChartCursorState,
+      createFreeState: createFreeChartCursorState,
+    }
+  },
 }
 
 /** Projects shared cursor state into one chart's plot and scales. */
@@ -199,84 +260,6 @@ export function resolveChartCursorFocus<
   return point ? (strategy?.group(points, { point }) ?? [point]) : []
 }
 
-/** Resolves built-in focus names without importing renderer or platform code. */
-export function resolveChartFocusStrategy<
-  TDatum,
-  TXValue extends ChartValue,
-  TYValue extends ChartValue,
->(
-  focus: ChartFocusMode<TDatum, TXValue, TYValue> | undefined,
-): ChartFocusStrategy<TDatum, TXValue, TYValue> | undefined {
-  if (focus === false) return undefined
-  if (typeof focus !== 'string') return focus
-  switch (focus) {
-    case 'nearest-x':
-      return focusNearestX
-    case 'nearest-y':
-      return focusNearestY
-    case 'group-x':
-      return focusX
-    case 'group-y':
-      return focusY
-    case 'nearest':
-      return undefined
-  }
-}
-
-/**
- * Resolves explicit pointer focus, or returns undefined for default nearest
- * focus. An empty array means the explicit strategy found no target. A points
- * array distinct from scene.points preserves presentation-point resolution.
- */
-export function resolveChartPointerFocus<
-  TDatum,
-  TXValue extends ChartValue,
-  TYValue extends ChartValue,
->(
-  scene: ChartScene<TDatum, TXValue, TYValue>,
-  focusMode: ChartFocusMode<TDatum, TXValue, TYValue> | undefined,
-  x: number,
-  y: number,
-  maxDistance: number,
-  points: readonly ChartPoint<TDatum, TXValue, TYValue>[] = scene.points,
-): readonly ChartPoint<TDatum, TXValue, TYValue>[] | undefined {
-  const strategy = resolveChartFocusStrategy(focusMode)
-  if (!strategy) return undefined
-  if (
-    points === scene.points &&
-    (strategy === focusNearestX ||
-      strategy === focusNearestY ||
-      strategy === focusX ||
-      strategy === focusY)
-  ) {
-    const contained = findContainingScenePoint(scene, x, y)
-    if (contained) {
-      return contained.point
-        ? strategy.group(points, { point: contained.point })
-        : []
-    }
-  }
-  return strategy.resolve(points, { x, y, maxDistance })
-}
-
-export function sameChartPointIdentity<
-  TDatum,
-  TXValue extends ChartValue,
-  TYValue extends ChartValue,
->(
-  left: ChartPoint<TDatum, TXValue, TYValue> | null,
-  right: ChartPoint<TDatum, TXValue, TYValue> | null,
-) {
-  return (
-    left === right ||
-    (left !== null &&
-      right !== null &&
-      left.key === right.key &&
-      left.markId === right.markId &&
-      left.datumIndex === right.datumIndex)
-  )
-}
-
 function cursorPointIdentity(
   point: Pick<ChartPoint, 'key' | 'markId' | 'datumIndex'>,
 ): ChartCursorPointIdentity {
@@ -301,171 +284,6 @@ function cursorPointFromIdentity<
     )
   }
   return undefined
-}
-
-export function restoreChartFocusPoint<
-  TDatum,
-  TXValue extends ChartValue,
-  TYValue extends ChartValue,
->(
-  points: readonly ChartPoint<TDatum, TXValue, TYValue>[],
-  previous: ChartPoint<TDatum, TXValue, TYValue>,
-) {
-  const matches = points.filter((point) => point.key === previous.key)
-  if (matches.length < 2) return matches[0] ?? null
-
-  const datumType = typeof previous.datum
-  const hasReferenceIdentity =
-    previous.datum !== null &&
-    (datumType === 'object' || datumType === 'function')
-  if (hasReferenceIdentity) {
-    const sameDatum = matches.find((point) => point.datum === previous.datum)
-    if (sameDatum) return sameDatum
-  }
-
-  return (
-    matches.find(
-      (point) =>
-        point.markId === previous.markId &&
-        Object.is(point.group, previous.group) &&
-        sameChartValue(point.xValue, previous.xValue) &&
-        sameChartValue(point.yValue, previous.yValue),
-    ) ??
-    matches.find(
-      (point) =>
-        point.markId === previous.markId &&
-        point.datumIndex === previous.datumIndex,
-    ) ??
-    matches[0] ??
-    null
-  )
-}
-
-export function chartPointFromNavigationOrder<
-  TDatum,
-  TXValue extends ChartValue,
-  TYValue extends ChartValue,
->(
-  points: readonly ChartPoint<TDatum, TXValue, TYValue>[],
-  current: ChartPoint<TDatum, TXValue, TYValue> | null,
-  key: string,
-): ChartPoint<TDatum, TXValue, TYValue> | null | undefined {
-  const currentIndex = current
-    ? points.findIndex((point) => sameChartPointIdentity(point, current))
-    : -1
-  let nextIndex: number
-  switch (key) {
-    case 'ArrowRight':
-    case 'ArrowDown':
-      nextIndex = Math.min(points.length - 1, currentIndex + 1)
-      break
-    case 'ArrowLeft':
-    case 'ArrowUp':
-      nextIndex = Math.max(0, currentIndex < 0 ? 0 : currentIndex - 1)
-      break
-    case 'Home':
-      nextIndex = 0
-      break
-    case 'End':
-      nextIndex = points.length - 1
-      break
-    default:
-      return undefined
-  }
-  return points[nextIndex] ?? null
-}
-
-export function chartPointFromSceneOrder<
-  TDatum,
-  TXValue extends ChartValue,
-  TYValue extends ChartValue,
->(
-  points: readonly ChartPoint<TDatum, TXValue, TYValue>[],
-  current: ChartPoint<TDatum, TXValue, TYValue> | null,
-  key: string,
-): ChartPoint<TDatum, TXValue, TYValue> | null | undefined {
-  const direction =
-    key === 'ArrowRight' || key === 'ArrowDown'
-      ? 1
-      : key === 'ArrowLeft' || key === 'ArrowUp'
-        ? -1
-        : key === 'Home'
-          ? 0
-          : key === 'End'
-            ? 2
-            : undefined
-  if (direction === undefined) return undefined
-  if (!points.length) return null
-
-  const currentIndex = current
-    ? points.findIndex((point) => sameChartPointIdentity(point, current))
-    : -1
-  if (!current || currentIndex < 0 || direction === 0 || direction === 2) {
-    return navigationExtreme(points, direction === 2)
-  }
-
-  let candidate: ChartPoint<TDatum, TXValue, TYValue> | null = null
-  let candidateIndex = -1
-  for (let index = 0; index < points.length; index += 1) {
-    const point = points[index]
-    if (!point) continue
-    const relative = compareNavigationPoints(
-      point,
-      index,
-      current,
-      currentIndex,
-    )
-    if ((direction > 0 && relative <= 0) || (direction < 0 && relative >= 0)) {
-      continue
-    }
-    if (
-      !candidate ||
-      direction *
-        compareNavigationPoints(point, index, candidate, candidateIndex) <
-        0
-    ) {
-      candidate = point
-      candidateIndex = index
-    }
-  }
-  return candidate ?? current
-}
-
-function navigationExtreme<
-  TDatum,
-  TXValue extends ChartValue,
-  TYValue extends ChartValue,
->(points: readonly ChartPoint<TDatum, TXValue, TYValue>[], maximum: boolean) {
-  let candidate = points[0] ?? null
-  let candidateIndex = 0
-  for (let index = 1; index < points.length; index += 1) {
-    const point = points[index]
-    if (!point || !candidate) continue
-    const comparison = compareNavigationPoints(
-      point,
-      index,
-      candidate,
-      candidateIndex,
-    )
-    if ((maximum && comparison > 0) || (!maximum && comparison < 0)) {
-      candidate = point
-      candidateIndex = index
-    }
-  }
-  return candidate
-}
-
-function compareNavigationPoints<
-  TDatum,
-  TXValue extends ChartValue,
-  TYValue extends ChartValue,
->(
-  left: ChartPoint<TDatum, TXValue, TYValue>,
-  leftIndex: number,
-  right: ChartPoint<TDatum, TXValue, TYValue>,
-  rightIndex: number,
-) {
-  return left.x - right.x || left.y - right.y || leftIndex - rightIndex
 }
 
 export function sameChartValue(left: unknown, right: unknown) {
