+++ packages/charts-core/src/index.ts
@@ -204,6 +204,7 @@ export type {
   ChartCursorBinding,
   ChartCursorController,
   ChartCursorCoordinates,
+  ChartCursorExtensionToken,
   ChartCursorPointIdentity,
   ChartCursorPresentation,
   ChartCursorState,
@@ -329,6 +330,8 @@ export type {
   SceneFocusGuideBand,
   SceneFocusGuideLabel,
   SceneFocusGuideMarker,
+  SceneFocusGuideResolveContext,
+  SceneFocusGuideResolver,
   SceneInteraction,
   SceneGroup,
   SceneLabel,
