+++ packages/charts-core/src/interaction.ts
@@ -1,0 +1,258 @@
+import { focusNearestX, focusNearestY, focusX, focusY } from './focus'
+import { findContainingScenePoint } from './nearest'
+import type {
+  ChartFocusMode,
+  ChartFocusStrategy,
+  ChartPoint,
+  ChartScene,
+  ChartValue,
+} from './types'
+
+/** Resolves built-in focus names without importing cursor or platform code. */
+export function resolveChartFocusStrategy<
+  TDatum,
+  TXValue extends ChartValue,
+  TYValue extends ChartValue,
+>(
+  focus: ChartFocusMode<TDatum, TXValue, TYValue> | undefined,
+): ChartFocusStrategy<TDatum, TXValue, TYValue> | undefined {
+  if (focus === false) return undefined
+  if (typeof focus !== 'string') return focus
+  switch (focus) {
+    case 'nearest-x':
+      return focusNearestX
+    case 'nearest-y':
+      return focusNearestY
+    case 'group-x':
+      return focusX
+    case 'group-y':
+      return focusY
+    case 'nearest':
+      return undefined
+  }
+}
+
+/**
+ * Resolves explicit pointer focus, or returns undefined for default nearest
+ * focus. An empty array means the explicit strategy found no target. A points
+ * array distinct from scene.points preserves presentation-point resolution.
+ */
+export function resolveChartPointerFocus<
+  TDatum,
+  TXValue extends ChartValue,
+  TYValue extends ChartValue,
+>(
+  scene: ChartScene<TDatum, TXValue, TYValue>,
+  focusMode: ChartFocusMode<TDatum, TXValue, TYValue> | undefined,
+  x: number,
+  y: number,
+  maxDistance: number,
+  points: readonly ChartPoint<TDatum, TXValue, TYValue>[] = scene.points,
+): readonly ChartPoint<TDatum, TXValue, TYValue>[] | undefined {
+  const strategy = resolveChartFocusStrategy(focusMode)
+  if (!strategy) return undefined
+  if (
+    points === scene.points &&
+    (strategy === focusNearestX ||
+      strategy === focusNearestY ||
+      strategy === focusX ||
+      strategy === focusY)
+  ) {
+    const contained = findContainingScenePoint(scene, x, y)
+    if (contained) {
+      return contained.point
+        ? strategy.group(points, { point: contained.point })
+        : []
+    }
+  }
+  return strategy.resolve(points, { x, y, maxDistance })
+}
+
+export function sameChartPointIdentity<
+  TDatum,
+  TXValue extends ChartValue,
+  TYValue extends ChartValue,
+>(
+  left: ChartPoint<TDatum, TXValue, TYValue> | null,
+  right: ChartPoint<TDatum, TXValue, TYValue> | null,
+) {
+  return (
+    left === right ||
+    (left !== null &&
+      right !== null &&
+      left.key === right.key &&
+      left.markId === right.markId &&
+      left.datumIndex === right.datumIndex)
+  )
+}
+
+export function restoreChartFocusPoint<
+  TDatum,
+  TXValue extends ChartValue,
+  TYValue extends ChartValue,
+>(
+  points: readonly ChartPoint<TDatum, TXValue, TYValue>[],
+  previous: ChartPoint<TDatum, TXValue, TYValue>,
+) {
+  const matches = points.filter((point) => point.key === previous.key)
+  if (matches.length < 2) return matches[0] ?? null
+
+  const datumType = typeof previous.datum
+  const hasReferenceIdentity =
+    previous.datum !== null &&
+    (datumType === 'object' || datumType === 'function')
+  if (hasReferenceIdentity) {
+    const sameDatum = matches.find((point) => point.datum === previous.datum)
+    if (sameDatum) return sameDatum
+  }
+
+  return (
+    matches.find(
+      (point) =>
+        point.markId === previous.markId &&
+        Object.is(point.group, previous.group) &&
+        sameChartValue(point.xValue, previous.xValue) &&
+        sameChartValue(point.yValue, previous.yValue),
+    ) ??
+    matches.find(
+      (point) =>
+        point.markId === previous.markId &&
+        point.datumIndex === previous.datumIndex,
+    ) ??
+    matches[0] ??
+    null
+  )
+}
+
+export function chartPointFromNavigationOrder<
+  TDatum,
+  TXValue extends ChartValue,
+  TYValue extends ChartValue,
+>(
+  points: readonly ChartPoint<TDatum, TXValue, TYValue>[],
+  current: ChartPoint<TDatum, TXValue, TYValue> | null,
+  key: string,
+): ChartPoint<TDatum, TXValue, TYValue> | null | undefined {
+  const currentIndex = current
+    ? points.findIndex((point) => sameChartPointIdentity(point, current))
+    : -1
+  let nextIndex: number
+  switch (key) {
+    case 'ArrowRight':
+    case 'ArrowDown':
+      nextIndex = Math.min(points.length - 1, currentIndex + 1)
+      break
+    case 'ArrowLeft':
+    case 'ArrowUp':
+      nextIndex = Math.max(0, currentIndex < 0 ? 0 : currentIndex - 1)
+      break
+    case 'Home':
+      nextIndex = 0
+      break
+    case 'End':
+      nextIndex = points.length - 1
+      break
+    default:
+      return undefined
+  }
+  return points[nextIndex] ?? null
+}
+
+export function chartPointFromSceneOrder<
+  TDatum,
+  TXValue extends ChartValue,
+  TYValue extends ChartValue,
+>(
+  points: readonly ChartPoint<TDatum, TXValue, TYValue>[],
+  current: ChartPoint<TDatum, TXValue, TYValue> | null,
+  key: string,
+): ChartPoint<TDatum, TXValue, TYValue> | null | undefined {
+  const direction =
+    key === 'ArrowRight' || key === 'ArrowDown'
+      ? 1
+      : key === 'ArrowLeft' || key === 'ArrowUp'
+        ? -1
+        : key === 'Home'
+          ? 0
+          : key === 'End'
+            ? 2
+            : undefined
+  if (direction === undefined) return undefined
+  if (!points.length) return null
+
+  const currentIndex = current
+    ? points.findIndex((point) => sameChartPointIdentity(point, current))
+    : -1
+  if (!current || currentIndex < 0 || direction === 0 || direction === 2) {
+    return navigationExtreme(points, direction === 2)
+  }
+
+  let candidate: ChartPoint<TDatum, TXValue, TYValue> | null = null
+  let candidateIndex = -1
+  for (let index = 0; index < points.length; index += 1) {
+    const point = points[index]
+    if (!point) continue
+    const relative = compareNavigationPoints(
+      point,
+      index,
+      current,
+      currentIndex,
+    )
+    if ((direction > 0 && relative <= 0) || (direction < 0 && relative >= 0)) {
+      continue
+    }
+    if (
+      !candidate ||
+      direction *
+        compareNavigationPoints(point, index, candidate, candidateIndex) <
+        0
+    ) {
+      candidate = point
+      candidateIndex = index
+    }
+  }
+  return candidate ?? current
+}
+
+function navigationExtreme<
+  TDatum,
+  TXValue extends ChartValue,
+  TYValue extends ChartValue,
+>(points: readonly ChartPoint<TDatum, TXValue, TYValue>[], maximum: boolean) {
+  let candidate = points[0] ?? null
+  let candidateIndex = 0
+  for (let index = 1; index < points.length; index += 1) {
+    const point = points[index]
+    if (!point || !candidate) continue
+    const comparison = compareNavigationPoints(
+      point,
+      index,
+      candidate,
+      candidateIndex,
+    )
+    if ((maximum && comparison > 0) || (!maximum && comparison < 0)) {
+      candidate = point
+      candidateIndex = index
+    }
+  }
+  return candidate
+}
+
+function compareNavigationPoints<
+  TDatum,
+  TXValue extends ChartValue,
+  TYValue extends ChartValue,
+>(
+  left: ChartPoint<TDatum, TXValue, TYValue>,
+  leftIndex: number,
+  right: ChartPoint<TDatum, TXValue, TYValue>,
+  rightIndex: number,
+) {
+  return left.x - right.x || left.y - right.y || leftIndex - rightIndex
+}
+
+function sameChartValue(left: ChartValue, right: ChartValue) {
+  return left instanceof Date && right instanceof Date
+    ? left.getTime() === right.getTime()
+    : Object.is(left, right)
+}
