+++ packages/charts-core/src/universal-types.ts
@@ -120,6 +120,7 @@ export type {
   ChartCursorBinding,
   ChartCursorController,
   ChartCursorCoordinates,
+  ChartCursorExtensionToken,
   ChartCursorPointIdentity,
   ChartCursorPresentation,
   ChartCursorState,
@@ -245,6 +246,8 @@ export type {
   SceneFocusGuideBand,
   SceneFocusGuideLabel,
   SceneFocusGuideMarker,
+  SceneFocusGuideResolveContext,
+  SceneFocusGuideResolver,
   SceneInteraction,
   SceneGroup,
   SceneLabel,
